requirements.txt
requests2==2.16.0
pandas==1.0.1
lxml==4.5.0
beautifulsoup4==4.8.2
selenium==3.141.0beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
idna==2.5
lxml==4.5.0
numpy==1.19.1
pandas==1.0.1
python-dateutil==2.8.1
pytz==2020.1
requests2==2.16.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.21.1
XlsxWriter==1.3.3
